services/shared/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agenticos-shared"
version = "0.1.0"
description = "AgenticOS shared library: settings, db, auth, audit, otel."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
dependencies = [
    "fastapi>=0.115,<1",
    "pydantic>=2.8,<3",
    "pydantic-settings>=2.4,<3",
    "sqlalchemy>=2.0.32,<3",
    "psycopg[binary]>=3.2,<4",
    "alembic>=1.13,<2",
    "redis>=5.0,<6",
    "nats-py>=2.7,<3",
    "structlog>=24.4,<25",
    "python-jose[cryptography]>=3.3,<4",
    "httpx>=0.27,<1",
    "opentelemetry-api>=1.27,<2",
    "opentelemetry-sdk>=1.27,<2",
    "opentelemetry-exporter-otlp>=1.27,<2",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.48b0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.24,<1",
    "httpx>=0.27,<1",
]

[tool.setuptools.packages.find]
where = ["src"]
